refactor(web): rename @stigmer/react-ui to @stigmer/agent-execution-ui
Align domain package name with ubiquitous language from the
agentic/agentexecution/v1/ bounded context. The package contains
exclusively AgentExecution-scoped components (streaming, tool calls,
HITL approvals, sub-agent delegation) — the new name communicates
this precisely and leaves clean namespace for future packages
(@stigmer/workflow-execution-ui, @stigmer/session-ui).

- Renamed directory _libs/domain/react-ui/ → agent-execution-ui/
- Updated all consumer imports (4 files)
- Updated build scripts, transpilePackages, workspace config
- Added **/dist/ to .prettierignore (pre-existing oversight)
- Recorded design decision with rationale and alternatives

Made-with: Cursor
