+++ client-apps/web/.prettierignore
@@ -1,6 +1,7 @@
 .next/
 out/
 build/
+**/dist/
 node_modules/
 *.d.ts
 next-env.d.ts
