+++ client-apps/web/_libs/domain/agent-execution-ui/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@stigmer/react-ui",
+  "name": "@stigmer/agent-execution-ui",
   "version": "0.0.0-dev",
   "description": "React components for Stigmer agent execution streaming, tool calls, and HITL approvals",
   "license": "Apache-2.0",
@@ -52,7 +52,7 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/stigmer/stigmer.git",
-    "directory": "client-apps/web/_libs/domain/react-ui"
+    "directory": "client-apps/web/_libs/domain/agent-execution-ui"
   },
   "dependencies": {
     "@stigmer/rpc-client": "*",
