+++ client-apps/web/_libs/domain/agent-execution-ui/src/execution/helpers.ts
