+++ client-apps/web/_libs/domain/agent-execution-ui/src/internal/ui/badge.tsx
